requirements.txt
pandas
xmltodict
lxml
python-dotenv
plotly
dash
matplotlib
seaborn
pyarrow
sqlite3
